requirements.txt
mysql-connector-python==8.2.0
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
wxPython==4.2.1